+++ TdlibKit/Models/BotCommands.swift
@@ -12,14 +12,14 @@ import Foundation
 public struct BotCommands: Codable {
 
     /// Bot's user identifier
-    public let botUserId: Int
+    public let botUserId: Int64
 
     /// List of bot commands
     public let commands: [BotCommand]
 
 
     public init(
-        botUserId: Int,
+        botUserId: Int64,
         commands: [BotCommand]
     ) {
         self.botUserId = botUserId
