+++ TdlibKit/Models/ChatStatisticsInviterInfo.swift
@@ -15,12 +15,12 @@ public struct ChatStatisticsInviterInfo: Codable {
     public let addedMemberCount: Int
 
     /// User identifier
-    public let userId: Int
+    public let userId: Int64
 
 
     public init(
         addedMemberCount: Int,
-        userId: Int
+        userId: Int64
     ) {
         self.addedMemberCount = addedMemberCount
         self.userId = userId
