+++ TdlibKit/Models/CreateCall.swift
@@ -14,17 +14,17 @@ public struct CreateCall: Codable {
     /// True, if a video call needs to be created
     public let isVideo: Bool?
 
-    /// Description of the call protocols supported by the application
+    /// The call protocols supported by the application
     public let `protocol`: CallProtocol?
 
     /// Identifier of the user to be called
-    public let userId: Int?
+    public let userId: Int64?
 
 
     public init(
         isVideo: Bool?,
         `protocol`: CallProtocol?,
-        userId: Int?
+        userId: Int64?
     ) {
         self.isVideo = isVideo
         self.`protocol` = `protocol`
