+++ TdlibKit/Models/CreateNewSecretChat.swift
@@ -12,10 +12,10 @@ import Foundation
 public struct CreateNewSecretChat: Codable {
 
     /// Identifier of the target user
-    public let userId: Int?
+    public let userId: Int64?
 
 
-    public init(userId: Int?) {
+    public init(userId: Int64?) {
         self.userId = userId
     }
 }
