+++ TdlibKit/Models/DeleteAllRevokedChatInviteLinks.swift
@@ -15,12 +15,12 @@ public struct DeleteAllRevokedChatInviteLinks: Codable {
     public let chatId: Int64?
 
     /// User identifier of a chat administrator, which links will be deleted. Must be an identifier of the current user for non-owner
-    public let creatorUserId: Int?
+    public let creatorUserId: Int64?
 
 
     public init(
         chatId: Int64?,
-        creatorUserId: Int?
+        creatorUserId: Int64?
     ) {
         self.chatId = chatId
         self.creatorUserId = creatorUserId
