+++ TdlibKit/Models/GetLoginUrlInfo.swift
@@ -12,7 +12,7 @@ import Foundation
 public struct GetLoginUrlInfo: Codable {
 
     /// Button identifier
-    public let buttonId: Int?
+    public let buttonId: Int64?
 
     /// Chat identifier of the message with the button
     public let chatId: Int64?
@@ -22,7 +22,7 @@ public struct GetLoginUrlInfo: Codable {
 
 
     public init(
-        buttonId: Int?,
+        buttonId: Int64?,
         chatId: Int64?,
         messageId: Int64?
     ) {
