+++ TdlibKit/Models/GetSupergroupMembers.swift
@@ -8,10 +8,10 @@
 import Foundation
 
 
-/// Returns information about members or banned users in a supergroup or channel. Can be used only if SupergroupFullInfo.can_get_members == true; additionally, administrator privileges may be required for some filters
+/// Returns information about members or banned users in a supergroup or channel. Can be used only if supergroupFullInfo.can_get_members == true; additionally, administrator privileges may be required for some filters
 public struct GetSupergroupMembers: Codable {
 
-    /// The type of users to return. By default, supergroupMembersFilterRecent
+    /// The type of users to return; pass null to use supergroupMembersFilterRecent
     public let filter: SupergroupMembersFilter?
 
     /// The maximum number of users be returned; up to 200
@@ -21,14 +21,14 @@ public struct GetSupergroupMembers: Codable {
     public let offset: Int?
 
     /// Identifier of the supergroup or channel
-    public let supergroupId: Int?
+    public let supergroupId: Int64?
 
 
     public init(
         filter: SupergroupMembersFilter?,
         limit: Int?,
         offset: Int?,
-        supergroupId: Int?
+        supergroupId: Int64?
     ) {
         self.filter = filter
         self.limit = limit
