+++ TdlibKit/Models/InviteGroupCallParticipants.swift
@@ -8,19 +8,19 @@
 import Foundation
 
 
-/// Invites users to an active group call. Sends a service message of type messageInviteToGroupCall for voice chats
+/// Invites users to an active group call. Sends a service message of type messageInviteToGroupCall for video chats
 public struct InviteGroupCallParticipants: Codable {
 
     /// Group call identifier
     public let groupCallId: Int?
 
     /// User identifiers. At most 10 users can be invited simultaneously
-    public let userIds: [Int]?
+    public let userIds: [Int64]?
 
 
     public init(
         groupCallId: Int?,
-        userIds: [Int]?
+        userIds: [Int64]?
     ) {
         self.groupCallId = groupCallId
         self.userIds = userIds
