+++ TdlibKit/Models/MessageSender.swift
@@ -53,10 +53,10 @@ public enum MessageSender: Codable {
 public struct MessageSenderUser: Codable {
 
     /// Identifier of the user that sent the message
-    public let userId: Int
+    public let userId: Int64
 
 
-    public init(userId: Int) {
+    public init(userId: Int64) {
         self.userId = userId
     }
 }
