+++ TdlibKit/Models/SetStickerSetThumbnail.swift
@@ -14,17 +14,17 @@ public struct SetStickerSetThumbnail: Codable {
     /// Sticker set name
     public let name: String?
 
-    /// Thumbnail to set in PNG or TGS format. Animated thumbnail must be set for animated sticker sets and only for them. Pass a zero InputFileId to delete the thumbnail
+    /// Thumbnail to set in PNG or TGS format; pass null to remove the sticker set thumbnail. Animated thumbnail must be set for animated sticker sets and only for them
     public let thumbnail: InputFile?
 
     /// Sticker set owner
-    public let userId: Int?
+    public let userId: Int64?
 
 
     public init(
         name: String?,
         thumbnail: InputFile?,
-        userId: Int?
+        userId: Int64?
     ) {
         self.name = name
         self.thumbnail = thumbnail
