+++ TdlibKit/Models/SetSupergroupUsername.swift
@@ -12,14 +12,14 @@ import Foundation
 public struct SetSupergroupUsername: Codable {
 
     /// Identifier of the supergroup or channel
-    public let supergroupId: Int?
+    public let supergroupId: Int64?
 
     /// New value of the username. Use an empty string to remove the username
     public let username: String?
 
 
     public init(
-        supergroupId: Int?,
+        supergroupId: Int64?,
         username: String?
     ) {
         self.supergroupId = supergroupId
