+++ TdlibKit/Models/Supergroup.swift
@@ -21,7 +21,7 @@ public struct Supergroup: Codable {
     public let hasLocation: Bool
 
     /// Supergroup or channel identifier
-    public let id: Int
+    public let id: Int64
 
     /// True, if the supergroup is a broadcast group, i.e. only administrators can send messages and there is no limit on number of members
     public let isBroadcastGroup: Bool
@@ -47,7 +47,7 @@ public struct Supergroup: Codable {
     /// If non-empty, contains a human-readable description of the reason why access to this supergroup or channel must be restricted
     public let restrictionReason: String
 
-    /// True, if messages sent to the channel should contain information about the sender. This field is only applicable to channels
+    /// True, if messages sent to the channel need to contain information about the sender. This field is only applicable to channels
     public let signMessages: Bool
 
     /// Status of the current user in the supergroup or channel; custom title will be always empty
@@ -61,7 +61,7 @@ public struct Supergroup: Codable {
         date: Int,
         hasLinkedChat: Bool,
         hasLocation: Bool,
-        id: Int,
+        id: Int64,
         isBroadcastGroup: Bool,
         isChannel: Bool,
         isFake: Bool,
